con-rs: extract shared consensus primitives from net-node
Introduce a top-level `con-rs` crate as a peer of `net-rs` and `sim-rs`,
holding the protocol pieces both implementations should agree on. This
is step 1 of a planned extraction: lift the pure modules verbatim, no
API redesign yet.

Lifted from net-node/src/consensus/leios:
- pipeline.rs: PipelinePhase, EbElection, PipelineConfig (CIP-0164 phase
  math).  Drops EbElection.eb_point — only used for tracing — to avoid
  pulling minicbor into con-rs.  Aggregation logging now formats slot +
  hash-prefix directly.
- wfa.rs: persistent committee allocation, NPV eligibility signature +
  Bernoulli win count, build_committee, expected_committee_size.
- aggregation.rs: record_vote / QuorumFormed.

Lifted from net-node/src/config.rs:
- CommitteeSelection (WfaLs / EveryoneVotes / StakeCentile) and its
  serde defaults; StakeEntry.  net-node still imports these from
  `crate::config` via `pub use con_rs::*`.

con-rs is sans-IO: deps are just serde / rand / blake2b_simd / tracing.
No tokio, no net-core.  Both consumers will reference it via path; sim-rs
adoption follows in a later commit.

Net-node delta: add con-rs path dep, replace the four module declarations
with `use con_rs::*`, drop the eb_point construction site.  voting.rs is
unchanged — `super::wfa` resolves through mod.rs's re-import — and is
deliberately left in net-node for now since it sends NetworkCommand and
belongs in step 2's sans-IO redesign.

Verified: cargo build, cargo test (28 con-rs + 544 net-rs workspace),
cargo clippy --all-targets (11 warnings, all pre-existing — no regression).
