+++ con-rs/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "con-rs"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+serde = { version = "1", features = ["derive"] }
+rand = "0.8"
+blake2b_simd = "1"
+tracing = "0.1"
