+++ net-rs/net-node/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+con-rs = { path = "../../con-rs" }
 net-core = { path = "../net-core" }
 tokio = { version = "1", features = ["rt-multi-thread", "macros", "time", "signal", "io-std", "io-util"] }
 clap = { version = "4", features = ["derive"] }
